fix IntegrationTest
